Вынес вкладку расписания в отдельный keepAlive виджет
